requirements.txt
Python 3.10+
Flask
PyTorch
OpenCV
NumPy
Pillow

pip install -r requirements.txt